+++ test_codecept/backendMock/local-mock.json
@@ -27,12 +27,13 @@
         },
         "waWorkflowApi": "http://localhost:8080",
         "documentsv2": {
-            "api": "http://localhost:8080/cases/documents"
+            "api": "http://localhost:8080/"
         },
         "prd": {
             "api": "http://localhost:8080",
             "commondataApi": "http://localhost:8080",
-            "locationApi": "http://localhost:8080"
+            "locationApi": "http://localhost:8080",
+            "judicialApi": "http://localhost:8080"
         },
         "em_anno_api": "http://localhost:8080",
         "idam": {
